fix: move CreateService to dedicated module under common/services
