+++ src/app/frontend/create/from/form/component.ts
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 import {Component, Input, OnInit} from '@angular/core';
+import {CreateService} from '../../../common/services/create/service';
 import {HistoryService} from '../../../common/services/global/history';
 import {NamespaceService} from '../../../common/services/global/namespace';
 import {overviewState} from '../../../overview/state';
-import {CreateService} from '../../service';
 
 // TODO
 @Component(
